requirements.txt
Django==3.1.1
djangorestframework==3.11.1
postgres==3.0.0
psycopg2-binary==2.8.6
djoser==2.0.5
drf-yasg==1.17.1
celery==4.4.7
redis==3.5.3
django-environ==0.4.5
